Contribuição Hacktoberfest - Orginização da Aplicação
