+++ AMS.Animais.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32825.248
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Animais", "Animais\Animais.csproj", "{580516BD-9C06-4878-A853-06EB97C76085}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{580516BD-9C06-4878-A853-06EB97C76085}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{580516BD-9C06-4878-A853-06EB97C76085}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{580516BD-9C06-4878-A853-06EB97C76085}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{580516BD-9C06-4878-A853-06EB97C76085}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0BC1CE6D-8EC0-40F0-8423-67CC160E066A}
+	EndGlobalSection
+EndGlobal
