+++ Animais/Animal.cs
@@ -1,0 +1,12 @@
+﻿namespace Animais
+{
+	public class Animal
+	{
+		public string Cor { get; set; } = default!;
+
+		protected string Nome { get; set; } = default!;
+
+		public string Especies { get; set; } = default!;
+
+	}
+}
