+++ Animais/Gato.cs
@@ -1,0 +1,7 @@
+﻿namespace Animais
+{
+	public class Gato : Animal
+	{
+		public string Raca { get; set; } = default!;
+	}
+}
