+++ Animais/Program.cs
@@ -1,0 +1,8 @@
+﻿using Animais;
+
+Cachorro c = new();
+c.GravarNome("Bilu", "Preto", 0.45, "Vira-lata");
+c.MostarDados();
+
+
+Console.ReadKey();
