--- Animal.cs
@@ -1,13 +1,0 @@
-namespace Animal
-{
-    public class Animal
-    {
-        public string Cor { get; set; }
-
-        protected string Nome { get; set; }
-
-        public string Especies { get; set; }
-        
-        console.writeline("Hello Bruno!");
-    }
-}
