--- Gato.cs
@@ -1,7 +1,0 @@
-namespace Animal
-{
-    public class Gato : Animal
-    {
-        public string Raca { get; set; }
-    }
-}
